Split tests in 3 files for readability / maintenance
